Delegate all types of retries to the caller
Initially the client had full-support for retries but was later removed
in favour of the caller handling them. However, we still left support
for retries in the client for the HTTP 429 case. However this is not
really needed in the client. All the client needs to do in this case is
block sending for the time specified the server. The caller can continue
to handle retries for this case like it already does for all other
cases. Another behavioral change this introduces is that it blocks
all workers from sending in case of 429 instead of just the worker
that received the 429 response.